docs(API): Publish the stopped count as an integer (no-changelog)
`stopMany` returns a loop counter, so the value can never be fractional.
`.int()` makes the generator emit `type: integer` instead of `type: number`.

+++ packages/@n8n/api-types/src/dto/executions/stopped-execution-public.dto.ts
@@ -18,7 +18,7 @@ export const stoppedExecutionPublicSchema = z.object({
 export class StoppedExecutionPublicDto extends Z.class(stoppedExecutionPublicSchema.shape) {}
 
 export const stoppedExecutionsPublicSchema = z.object({
-	stopped: z.number().openapi(stoppedCountOpenApi),
+	stopped: z.number().int().openapi(stoppedCountOpenApi),
 });
 
 export class StoppedExecutionsPublicDto extends Z.class(stoppedExecutionsPublicSchema.shape) {}
